LayoutTests/imported/w3c/web-platform-tests/css/css-sizing/percentage-height-replaced-content-in-auto-cb-expected.txt

Test passes if there is a green square below.



FAIL img 1 assert_equals:
<img id="img" data-expected-height="100" style="width: 100px; height: 100%" src="/css/support/60x60-green.png?0.34388370032986304">
height expected 100 but got 0
